Add Notification Service, Persist Shopping-List Updates to LocalStorage, Enhance UI, Code Refactoring
- Created new 'AlertNotification' component, which gives feedback to the user about actions whose result is not immediately obvious in the UI
   - save recipe,
   - auth failure,
   - copying ingredients from recipe to shopping list
   - auto logout due to token expiration
  The old 'Alert' component - which was previously loaded dynamically used Angular's ComponentResolverFactory - has been removed, and the ComponentResolverFactory is now being used to dynamically load the new AlertNotification Component
- Changes to the shopping list are now persisted to LocalStorage, and pulled upon browser refresh.
  The idea was to make the process of updating a shopping list in a real-life scenario as quick as possible.
  To accomodate this, a new shopping-list effects class was added, which will listen to any of the shopping-list CRUD actions, then save the updated shopping-list state to LocalStorage.
- Other changes (among UI changes) include:
   - updating access modifiers on class props and functions,
   - ease-in/out animation when recipe or shopping-list are updated,
   - custom scrollbar,
   - reading strings from a constants file instead of using literals,
   - adding a new font,
   - consistent button sizing,
   - addition of icons to notification messages,
   - adding a suggestion to load recipes from database (or else add new recipes) when recipe list is empty,
   - tooltips on disabled buttons and pieces of text which could be clipped in their containers when the browser is zoomed in

+++ course-tutorials/course-app/prj-start/src/app/recipes/recipe-detail/recipe-detail.component.ts
@@ -1,8 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { map, switchMap } from 'rxjs/operators';
 
+import { Constants } from '../../shared/constants';
 import { Recipe } from '../recipe.model';
 import * as fromApp from '../../store/app.reducer';
 import * as RecipeActions from '../store/recipe.actions';
@@ -13,17 +14,24 @@ import * as ShoppingListActions from '../../shopping-list/store/shopping-list.ac
   templateUrl: './recipe-detail.component.html',
   styleUrls: ['./recipe-detail.component.css']
 })
-export class RecipeDetailComponent implements OnInit {
-  recipe: Recipe;
-  id: number;
+export class RecipeDetailComponent {
+  private _recipe: Recipe;
+  private _id: number;
+  constants = Constants;
+
+  get id(): number { return this._id; }
+  set id(e: number) { this._id = e; }
+  get recipe(): Recipe { return this._recipe; }
+  set recipe(e: Recipe) { this._recipe = e; }
 
   constructor(
     private router: Router,
     private route: ActivatedRoute,
     private store: Store<fromApp.AppState>
-  ) {}
-
-  ngOnInit() {
+  ) {
+    this.router.routeReuseStrategy.shouldReuseRoute = () => {
+      return false;
+    };
     this.route.params
       .pipe(
         map((params: Params) => {
@@ -40,22 +48,34 @@ export class RecipeDetailComponent implements OnInit {
         })
       )
       .subscribe(recipe => {
+        if (!recipe) {
+          return this.router.navigateByUrl('/recipes');
+        }
         this.recipe = recipe;
       });
+  }
+
+  isDisabled(recipe: Recipe): boolean {
+    return recipe.ingredients.length === 0;
   }
 
   onEditRecipe() {
     this.router.navigate(['edit'], { relativeTo: this.route });
   }
 
-  onAddToShoppingList() {
+  onAddToShoppingList(recipe: Recipe) {
+    if (this.isDisabled(recipe)) {
+      return;
+    }
     this.store.dispatch(
       new ShoppingListActions.AddIngredients(this.recipe.ingredients)
     );
   }
 
   onDeleteRecipe() {
-    this.store.dispatch(new RecipeActions.DeleteRecipe(this.id));
-    this.router.navigate(['/recipes']);
+    if (confirm(this.constants.confirmDeleteRecipeLabel)) {
+      this.store.dispatch(new RecipeActions.DeleteRecipe(this.id));
+      this.router.navigate(['/recipes']);
+    }
   }
 }
